// src/modules/monitor/log/api/types.ts
/*
 * @Author: outsider 515885633@qq.com
 * @Date: 2023-01-29
 * @LastEditors: outsider 515885633@qq.com
 * @FilePath: \vue-element-plus-admin\src\api\log\types.ts
 * @Description:
 *
 * Copyright (c) 2023 by outsider 515885633@qq.com, All Rights Reserved.
 */
export type LoggerType = {
  line: number
  id: string
  username: string
  method: string
  browser: string
  level: string
  message: string
}